refactor(doctor): unify health checks on the per-step model
Remove the aggregate report functions (`pourpoint_coverage_report`,
`aoi_health_report`, `grid_validation_report`) that duplicated the per-step
health checks `doctor` actually runs. They were a second, unused
implementation kept alive only by tests -- and `pourpoint_coverage_report`
had drifted, still emitting `no raster` for off-grid basins that `doctor`
suppresses. `run_health_checks` is now the single source of truth; their
tests select a single check (e.g. `run_health_checks(db, [ds],
['pourpoints'])`) instead.

Merge the three near-identical grid-check helpers (`_cog_grid_issues`,
`_cog_grid_findings`, `_full_grid_findings`) into one `_raster_grid_findings`
covering COGs, zone layers, and the nodata mask, enumerated via a single
`_grid_target_rasters` generator.

Extract the doctor machinery out of `diagnostics.py` (925 -> 310 lines) into a
new `snowtool.snowdb.health_checks` module; `diagnostics` now holds only the
read-only report builders. `doctor`'s startup announcement derives its check
list from the registry, and `_pourpoints_steps` computes each triplet once.

No change to CLI behavior or output. Released as v0.5.1.
