+++ src/snowtool/snowdb/diagnostics.py
@@ -14,24 +14,15 @@ import math
 
 from dataclasses import dataclass
 from datetime import date, timedelta
-from functools import partial
 from typing import TYPE_CHECKING
 
-from snowtool.exceptions import QueryParameterError, UnknownHealthCheckError
-from snowtool.snowdb import issues as issues_mod
-from snowtool.snowdb import triplet_naming
+from snowtool.exceptions import QueryParameterError
 from snowtool.snowdb.grid import grid_extent
-from snowtool.snowdb.progress import NULL_PROGRESS
-from snowtool.snowdb.query import DateRangeQuery
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterator, Sequence
-    from pathlib import Path
-
     from snowtool.snowdb.dataset import Dataset, DatasetArtifacts
     from snowtool.snowdb.db import SnowDb
     from snowtool.snowdb.grid import Extent
-    from snowtool.snowdb.progress import ProgressReporter
 
 
 @dataclass(frozen=True)
@@ -114,189 +105,6 @@ def missing_dates(
     return [
         d for i in range(n_days) if (d := start + timedelta(days=i)) not in ingested
     ]
-
-
-# --- report builders (read-only; the `dataset`/`doctor` commands render these) --
-#
-# A ``Finding`` is one row of ``snowtool doctor`` output: the flat, uniform
-# ``check``/``dataset``/``target``/``issue`` dict the CLI renders. The four
-# doctor-only checks below compose their ``target``/``issue`` prose *where the
-# condition is detected* and return these rows directly -- there is no separate
-# typed-intermediate tier for the CLI to re-flatten. The facts each row encodes
-# are pinned by ``tests/snowdb/test_report_diagnostics.py``.
-
-type Finding = dict[str, str]
-
-
-def _finding(check: str, dataset: str, target: str, issue: str) -> Finding:
-    return {'check': check, 'dataset': dataset, 'target': target, 'issue': issue}
-
-
-def _findings_from_issues(
-    check: str,
-    dataset: str,
-    target: str,
-    issue_list: Sequence[issues_mod.Issue],
-) -> list[Finding]:
-    """Render a batch of typed :class:`~snowtool.snowdb.issues.Issue` to findings
-    sharing one ``check``/``dataset``/``target``."""
-    return [_finding(check, dataset, target, i.message) for i in issue_list]
-
-
-def completeness_report(
-    dataset: Dataset,
-    *,
-    start: date | None = None,
-    end: date | None = None,
-) -> list[Finding]:
-    """``dates`` findings: ingested dates (optionally within ``start``/``end``)
-    missing one or more of their dataset's variables."""
-    window = DateRangeQuery(start_date=start, end_date=end)
-    findings: list[Finding] = []
-    for d in window.select(dataset.available_dates()):
-        findings.extend(_date_completeness(dataset, d))
-    return findings
-
-
-def _date_completeness(dataset: Dataset, d: date) -> list[Finding]:
-    """The ``dates`` finding (if any) for a single ingested date ``d``."""
-    unresolved = dataset.unresolved_variables(d)
-    if not unresolved:
-        return []
-    return [
-        _finding(
-            'dates',
-            dataset.spec.name,
-            d.isoformat(),
-            f'missing {", ".join(sorted(unresolved))}',
-        ),
-    ]
-
-
-def missing_artifacts(dataset: Dataset) -> list[str]:
-    """The dataset's expected on-disk artifacts that are absent.
-
-    Every configured zone layer (terrain, land cover, ...) is expected --
-    ``snowtool init`` builds each from its default source -- so a missing one is a
-    finding. An incomplete zone-layer set names the specific layer files that are
-    absent (``terrain (elevation.tif, aspect_majority.tif)``) so the finding is
-    actionable, not just the provider name.
-    """
-    artifacts = dataset.artifact_status()
-    missing: list[str] = []
-    for name, present in artifacts.zone_layers.items():
-        if present:
-            continue
-        absent = ', '.join(
-            layer.filename for layer in dataset.zones[name].missing_layers()
-        )
-        missing.append(f'{name} ({absent})' if absent else name)
-    if not artifacts.cogs:
-        missing.append('cogs')
-    if not artifacts.aoi_rasters:
-        missing.append('aoi-rasters')
-    return missing
-
-
-def stale_format_zone_layers(dataset: Dataset) -> list[Finding]:
-    """``files`` findings: built zone-layer sets stamped with an out-of-date
-    on-disk format version.
-
-    Skips sets that are not built (``missing_artifacts`` already reports those);
-    a built set whose stamped version differs from the provider's current one --
-    including a missing/legacy tag (stored ``None``) -- is flagged for a rebuild.
-    The ``target`` is the provider name; the ``issue`` names the stored vs.
-    current versions.
-    """
-    name = dataset.spec.name
-    findings: list[Finding] = []
-    for provider_name, zone_set in dataset.zones.items():
-        if zone_set.format_is_current() is False:
-            findings.append(
-                _finding(
-                    'files',
-                    name,
-                    provider_name,
-                    f'stale zone-layer format (stored '
-                    f'{zone_set.stored_format_version()} != '
-                    f'current {zone_set.format_version})',
-                ),
-            )
-    return findings
-
-
-def pourpoint_coverage_report(snowdb: SnowDb, dataset: Dataset) -> list[Finding]:
-    """``pourpoints`` findings for how a dataset's grid + burned rasters line up
-    with the stored pourpoints.
-
-    ``no raster``/``orphan raster`` are about which AOI *rasters* exist;
-    ``partial coverage``/``no coverage`` are the geometric coverage of each
-    pourpoint's basin by the dataset's grid (partial = basin spills outside it,
-    no coverage = basin entirely off-grid). A fully-covered, rasterized
-    pourpoint yields no finding. The ``target`` is the station triplet.
-    """
-    from snowtool.snowdb.coverage import Coverage, dataset_coverage
-
-    name = dataset.spec.name
-    triplets = snowdb.pourpoint_triplets()
-    rasterized = dataset.aoi_raster_triplets()
-    # Coverage is computed live from each stored basin (this is validation -- it
-    # must not trust the derived index, which it exists to catch drift in).
-    partial: list[str] = []
-    uncovered: list[str] = []
-    domain = dataset.coverage_domain
-    for pourpoint in snowdb.pourpoints():
-        match dataset_coverage(pourpoint, domain):
-            case Coverage.PARTIAL:
-                partial.append(pourpoint.station_triplet)
-            case Coverage.NONE:
-                uncovered.append(pourpoint.station_triplet)
-            case _:
-                pass
-    return [
-        _finding('pourpoints', name, triplet, issue)
-        for issue, triplet_set in (
-            ('no raster', triplets - rasterized),
-            ('orphan raster', rasterized - triplets),
-            ('partial coverage', set(partial)),
-            ('no coverage', set(uncovered)),
-        )
-        for triplet in sorted(triplet_set)
-    ]
-
-
-def aoi_health_report(dataset: Dataset) -> list[Finding]:
-    """``pourpoints`` findings for burned AOI rasters that won't read cleanly.
-
-    Opens each AOI raster; a read failure, a missing tile-bbox tag, a grid
-    mismatch, or an all-zero (empty) raster becomes a finding whose ``target``
-    is the station triplet and whose ``issue`` describes the fault. Freshness
-    is not checked here (no registry is available to resolve an expected hash
-    against) -- see :func:`_aoi_raster_findings` for the doctor step that adds it.
-    """
-    findings: list[Finding] = []
-    for path in dataset.aoi_raster_paths():
-        findings.extend(_aoi_raster_findings(dataset, path.stem, path, None))
-    return findings
-
-
-def _aoi_raster_findings(
-    dataset: Dataset,
-    stem: str,
-    path: Path,
-    expected_hash: str | None,
-) -> list[Finding]:
-    """The ``pourpoints`` findings for one burned AOI raster.
-
-    Delegates to :func:`~snowtool.snowdb.aoi_raster.aoi_raster_issues` -- the
-    shared structure/grid/freshness check -- and renders its typed
-    :class:`~snowtool.snowdb.issues.Issue` list to findings.
-    """
-    from snowtool.snowdb.aoi_raster import aoi_raster_issues
-
-    triplet = triplet_naming.stem_to_triplet(stem)
-    found = aoi_raster_issues(path, grid=dataset.grid, expected_hash=expected_hash)
-    return _findings_from_issues('pourpoints', dataset.spec.name, triplet, found)
 
 
 @dataclass(frozen=True)
@@ -500,426 +308,3 @@ def dataset_info_report(snowdb: SnowDb, dataset: Dataset) -> DatasetInfoReport:
             for name in dataset.zones
         },
     )
-
-
-def _iter_cog_paths(dataset: Dataset) -> Iterator[Path]:
-    """Every ingested COG on disk, ascending by date then filename.
-
-    Lists each ``cogs/<date>/`` directory once and yields all its ``.tif`` files,
-    so the grid check can validate *every* data file's header (not just a
-    representative), and tolerate a stray/duplicate COG a per-variable resolve
-    would raise on.
-    """
-    for d in dataset.available_dates():
-        yield from sorted(dataset.date_dir(d).glob('*.tif'))
-
-
-def _grid_declaration_issues(dataset: Dataset) -> list[str]:
-    """The declaration-only grid problems (no raster I/O).
-
-    An ingester with no variables has nothing to write -- almost certainly a
-    misconfiguration. (The reverse -- variables but no ingester -- is *not*
-    flagged: that is a valid read-only/derived dataset, populated out of band.)
-    A deeper variables-vs-ingester check (the ingester's *required* keys being a
-    subset of those declared) would need the ``Ingester`` protocol to expose its
-    expected keys; that is left as a follow-up.
-    """
-    spec = dataset.spec
-    if spec.ingester is not None and not spec.variables:
-        return ['has an ingester but declares no variables']
-    return []
-
-
-def _cog_grid_issues(dataset: Dataset, cog: Path) -> list[str]:
-    """Shape + transform problems for one COG against the declared grid.
-
-    Opens ``cog``'s header and checks its dimensions and transform against the
-    dataset's declared grid, catching a config that has drifted from the real
-    rasters (or a file ingested onto a different lattice). Delegates the
-    comparison to :func:`snowtool.snowdb.issues.grid_issues` -- the same check
-    used for every other grid-bound artifact -- and renders each returned
-    ``Issue`` back to a string so callers keep their ``list[str]`` shape.
-    """
-    import rasterio
-
-    grid = dataset.spec.grid_params
-    declared = dataset.grid.base_grid.transform
-    with rasterio.open(cog) as src:
-        actual = src.transform
-        width, height = src.width, src.height
-    found = issues_mod.grid_issues(
-        declared_transform=declared,
-        actual_transform=actual,
-        declared_shape=(grid.rows, grid.cols),
-        actual_shape=(height, width),
-    )
-    return [issue.message for issue in found]
-
-
-def grid_validation_report(dataset: Dataset) -> list[str]:
-    """Declaration-vs-reality grid problems (empty == consistent).
-
-    The declaration check (:func:`_grid_declaration_issues`) plus the shape +
-    transform check (:func:`_cog_grid_issues`) run against **every** ingested
-    COG, so a lattice drift on any date is caught, not just the first. The
-    ``doctor`` sweep enumerates the same work one COG at a time for progress
-    (see :func:`_grid_steps`); this aggregate is for direct callers/tests.
-    """
-    issues = _grid_declaration_issues(dataset)
-    for cog in _iter_cog_paths(dataset):
-        issues.extend(_cog_grid_issues(dataset, cog))
-    return issues
-
-
-# --- doctor: enumerable health-check steps -----------------------------------
-#
-# `run_health_checks` shows one progress increment per unit of work, so a check
-# is not a single ``(snowdb, dataset) -> findings`` call but an *enumeration* of
-# `CheckStep`s: a live-progress label plus a deferred `run`. Enumeration only
-# lists directories (cheap); each `run` does the expensive open/parse. Listing
-# every step across all datasets up front lets the reporter set an exact total
-# and announce each unit (a COG, an AOI raster, a date) as it runs.
-
-
-@dataclass(frozen=True)
-class CheckStep:
-    """One unit of ``doctor`` work: a progress ``label`` and a deferred ``run``
-    that produces its findings when executed."""
-
-    label: str
-    run: Callable[[], list[Finding]]
-
-
-def _grid_declaration_findings(dataset: Dataset, name: str) -> list[Finding]:
-    return [
-        _finding('grid', name, '', issue) for issue in _grid_declaration_issues(dataset)
-    ]
-
-
-def _cog_grid_findings(
-    dataset: Dataset,
-    name: str,
-    cog: Path,
-    target: str,
-) -> list[Finding]:
-    return [
-        _finding('grid', name, target, issue)
-        for issue in _cog_grid_issues(dataset, cog)
-    ]
-
-
-def _full_grid_findings(
-    dataset: Dataset,
-    name: str,
-    path: Path,
-    target: str,
-) -> list[Finding]:
-    """The ``grid`` findings for one full-grid-sized raster (a zone layer or the
-    nodata mask), against the dataset's declared grid.
-
-    Unlike a COG (also full-grid) or an AOI raster (a windowed crop, checked
-    under ``pourpoints``), this is a plain full-grid comparison: the file's own
-    transform/shape against ``dataset.spec.grid_params`` and
-    ``dataset.grid.base_grid.transform``.
-    """
-    import rasterio
-
-    grid = dataset.spec.grid_params
-    with rasterio.open(path) as src:
-        actual_transform = src.transform
-        actual_shape = (src.height, src.width)
-    found = issues_mod.grid_issues(
-        declared_transform=dataset.grid.base_grid.transform,
-        actual_transform=actual_transform,
-        declared_shape=(grid.rows, grid.cols),
-        actual_shape=actual_shape,
-    )
-    return _findings_from_issues('grid', name, target, found)
-
-
-def _grid_steps(_snowdb: SnowDb, dataset: Dataset) -> list[CheckStep]:
-    name = dataset.spec.name
-    steps = [
-        CheckStep(
-            f'{name} grid: declaration',
-            partial(_grid_declaration_findings, dataset, name),
-        ),
-    ]
-    # One step per COG so every data file's header is validated (not just a
-    # representative), each advancing the bar under its own label.
-    for cog in _iter_cog_paths(dataset):
-        target = f'{cog.parent.name}/{cog.name}'
-        steps.append(
-            CheckStep(
-                f'{name} grid: {target}',
-                partial(_cog_grid_findings, dataset, name, cog, target),
-            ),
-        )
-    # One step per present zone-layer file (terrain: elevation/aspect, land
-    # cover: forest, ...) -- a missing layer is the `files` check's concern, not
-    # grid, so only files that actually exist on disk are validated here.
-    for provider_name, zone_set in dataset.zones.items():
-        for layer in zone_set.layers:
-            path = zone_set.layer_path(layer)
-            if not path.is_file():
-                continue
-            target = f'{provider_name}/{layer.filename}'
-            steps.append(
-                CheckStep(
-                    f'{name} grid: {target}',
-                    partial(_full_grid_findings, dataset, name, path, target),
-                ),
-            )
-    # The dataset's nodata mask (when configured) is also a full-grid raster.
-    if dataset.nodata_mask is not None and dataset.nodata_mask.is_file():
-        steps.append(
-            CheckStep(
-                f'{name} grid: nodata-mask.tif',
-                partial(
-                    _full_grid_findings,
-                    dataset,
-                    name,
-                    dataset.nodata_mask,
-                    'nodata-mask.tif',
-                ),
-            ),
-        )
-    return steps
-
-
-def _dates_steps(_snowdb: SnowDb, dataset: Dataset) -> list[CheckStep]:
-    name = dataset.spec.name
-    return [
-        CheckStep(
-            f'{name} dates: {d.isoformat()}',
-            partial(_date_completeness, dataset, d),
-        )
-        for d in dataset.available_dates()
-    ]
-
-
-def _missing_artifact_findings(dataset: Dataset, name: str) -> list[Finding]:
-    return [
-        _finding('files', name, artifact, 'missing')
-        for artifact in missing_artifacts(dataset)
-    ]
-
-
-def _files_steps(_snowdb: SnowDb, dataset: Dataset) -> list[CheckStep]:
-    name = dataset.spec.name
-    return [
-        CheckStep(
-            f'{name} files: artifacts',
-            partial(_missing_artifact_findings, dataset, name),
-        ),
-        CheckStep(
-            f'{name} files: zone-layer formats',
-            partial(stale_format_zone_layers, dataset),
-        ),
-    ]
-
-
-def _pourpoint_coverage_findings(
-    dataset: Dataset,
-    record_path: Path,
-    triplet: str,
-    rasterized: frozenset[str],
-) -> list[Finding]:
-    """Coverage findings for one pourpoint basin (parses + reprojects it).
-
-    This is the per-basin unit of the coverage scan -- the expensive part
-    (parsing the record and reprojecting the basin into the grid CRS), one
-    pourpoint at a time so ``doctor``'s bar advances through a large registry
-    instead of stalling on a single monolithic step.
-
-    An off-grid basin (``NONE``) cannot be rasterized -- ``rasterize_aoi`` refuses
-    it and the batch path skips it -- so a *missing* raster is expected and only
-    ``no coverage`` is reported (a *stray* all-zero raster is still caught by the
-    per-raster AOI-validation steps). A ``PARTIAL`` basin can and should be
-    rasterized, so ``no raster`` stands there. Issue order (``no raster`` before
-    ``partial coverage``) matches the collapsed-row order the tests pin.
-    """
-    from snowtool.snowdb.coverage import Coverage, dataset_coverage
-    from snowtool.snowdb.pourpoint import Pourpoint
-
-    name = dataset.spec.name
-    coverage = dataset_coverage(
-        Pourpoint.from_basin_record(record_path),
-        dataset.coverage_domain,
-    )
-    if coverage is Coverage.NONE:
-        return _findings_from_issues(
-            'pourpoints',
-            name,
-            triplet,
-            [issues_mod.NoCoverage()],
-        )
-    found: list[issues_mod.Issue] = []
-    if triplet not in rasterized:
-        found.append(issues_mod.NoRaster())
-    if coverage is Coverage.PARTIAL:
-        found.append(issues_mod.PartialCoverage())
-    return _findings_from_issues('pourpoints', name, triplet, found)
-
-
-def _orphan_raster_findings(
-    name: str,
-    rasterized: frozenset[str],
-    triplets: set[str],
-) -> list[Finding]:
-    """``orphan raster`` findings: burned rasters with no backing record."""
-    findings: list[Finding] = []
-    for triplet in sorted(rasterized - triplets):
-        findings.extend(
-            _findings_from_issues(
-                'pourpoints',
-                name,
-                triplet,
-                [issues_mod.OrphanArtifact()],
-            ),
-        )
-    return findings
-
-
-def _pourpoints_steps(snowdb: SnowDb, dataset: Dataset) -> list[CheckStep]:
-    name = dataset.spec.name
-    # Cheap up front (filenames + a glob); the per-basin reprojection is deferred
-    # into one step each so the progress total is exact and the bar advances.
-    rasterized = frozenset(dataset.aoi_raster_triplets())
-    steps = [
-        CheckStep(
-            f'{name} pourpoints: coverage {triplet_naming.stem_to_triplet(p.stem)}',
-            partial(
-                _pourpoint_coverage_findings,
-                dataset,
-                p,
-                triplet_naming.stem_to_triplet(p.stem),
-                rasterized,
-            ),
-        )
-        for p in snowdb.pourpoint_paths()
-    ]
-    steps.append(
-        CheckStep(
-            f'{name} pourpoints: orphan rasters',
-            partial(
-                _orphan_raster_findings,
-                name,
-                rasterized,
-                snowdb.pourpoint_triplets(),
-            ),
-        ),
-    )
-    for path in dataset.aoi_raster_paths():
-        triplet = triplet_naming.stem_to_triplet(path.stem)
-        steps.append(
-            CheckStep(
-                f'{name} pourpoints: AOI validation {triplet}',
-                # Defer both the record parse (expected hash) and the raster read
-                # into the step body -- computing them here, during enumeration,
-                # parses every basin record before the progress bar opens.
-                partial(_aoi_raster_step_findings, snowdb, dataset, path),
-            ),
-        )
-    return steps
-
-
-def _aoi_raster_step_findings(
-    snowdb: SnowDb,
-    dataset: Dataset,
-    path: Path,
-) -> list[Finding]:
-    """A doctor AOI-validation step: resolve the expected hash from the registry
-    (deferred), then run the shared AOI check."""
-    triplet = triplet_naming.stem_to_triplet(path.stem)
-    expected_hash = _expected_aoi_hash(snowdb, dataset, triplet)
-    return _aoi_raster_findings(dataset, path.stem, path, expected_hash)
-
-
-def _expected_aoi_hash(
-    snowdb: SnowDb,
-    dataset: Dataset,
-    triplet: str,
-) -> str | None:
-    """The AOI provenance hash a raster for ``triplet`` should carry, or ``None``
-    when there is no backing basin record to compare against (an orphan raster,
-    reported separately by :func:`_orphan_raster_findings`)."""
-    from snowtool.snowdb.aoi_raster import aoi_provenance
-    from snowtool.snowdb.pourpoint import Pourpoint
-
-    record_path = snowdb.pourpoint_record_path(triplet)
-    if not record_path.is_file():
-        return None
-    pourpoint = Pourpoint.from_basin_record(record_path)
-    return aoi_provenance(pourpoint.geometry_hash, dataset.nodata_mask_hash)
-
-
-# Order is the ``doctor`` output/CLI-help order.
-HEALTH_CHECKS: dict[str, Callable[[SnowDb, Dataset], list[CheckStep]]] = {
-    'grid': _grid_steps,
-    'dates': _dates_steps,
-    'files': _files_steps,
-    'pourpoints': _pourpoints_steps,
-}
-
-# The valid check names, in ``doctor``'s output/default-sweep order.
-HEALTH_CHECK_NAMES: tuple[str, ...] = tuple(HEALTH_CHECKS)
-
-
-def run_health_checks(
-    snowdb: SnowDb,
-    datasets: Sequence[Dataset],
-    checks: Sequence[str],
-    *,
-    progress: ProgressReporter = NULL_PROGRESS,
-) -> list[Finding]:
-    """Sweep ``checks`` across ``datasets``, returning the flat finding rows
-    ``snowtool doctor`` renders (empty means healthy).
-
-    Resolves ``checks`` first: an unknown name raises
-    :class:`~snowtool.exceptions.UnknownHealthCheckError`, duplicates are
-    dropped (order-preserving), and an empty selection defaults to every check
-    in :data:`HEALTH_CHECK_NAMES`. Then enumerates every check's steps (one per
-    COG, AOI raster, date, ...) across all datasets up front so ``progress`` has
-    an exact total, and runs each step, naming it on the bar and advancing one
-    tick per step. Findings sharing a ``(check, dataset, target)`` are collapsed
-    onto one row (:func:`_collapse_by_target`).
-    """
-    unknown = sorted(set(checks) - set(HEALTH_CHECK_NAMES))
-    if unknown:
-        raise UnknownHealthCheckError(
-            f'Unknown check(s): {", ".join(unknown)}. '
-            f'Known checks: {", ".join(HEALTH_CHECK_NAMES)}.',
-        )
-    selected = list(dict.fromkeys(checks)) if checks else list(HEALTH_CHECK_NAMES)
-
-    steps: list[CheckStep] = []
-    for dataset in datasets:
-        for check in selected:
-            steps.extend(HEALTH_CHECKS[check](snowdb, dataset))
-
-    findings: list[Finding] = []
-    with progress.track('doctor', total=len(steps)) as task:
-        for step in steps:
-            task.describe(step.label)
-            findings.extend(step.run())
-            task.advance()
-    return _collapse_by_target(findings)
-
-
-def _collapse_by_target(findings: Sequence[Finding]) -> list[Finding]:
-    """Roll findings that share a ``(check, dataset, target)`` onto one row.
-
-    A target that trips several issues (e.g. a basin that is both ``no raster``
-    and ``partial coverage``) becomes a single row whose ``issue`` joins them with
-    ``'; '`` -- so ``doctor`` never spreads one target across multiple lines --
-    preserving first-seen order both across rows and within the joined issue.
-    """
-    collapsed: dict[tuple[str, str, str], Finding] = {}
-    for finding in findings:
-        key = (finding['check'], finding['dataset'], finding['target'])
-        if (existing := collapsed.get(key)) is not None:
-            existing['issue'] += f'; {finding["issue"]}'
-        else:
-            collapsed[key] = dict(finding)
-    return list(collapsed.values())
